refactor item
