+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import Faqs from '@/components/landing/FaqList';
 import MainInfoBlock from '@/components/landing/MainInfoBlock';
 import MotivationBlock from '@/components/landing/MotivationBlock';
-import PartnerBlock from '@/components/landing/PartnerBlock';
+import PartnerBlock from '@/components/landing/partner/PartnerBlock';
 import Timeline from '@/components/landing/ShortTimeline';
 import LongTimeline from '@/components/landing/Timeline';
 import WaitingForBlock from '@/components/landing/WaitingForBlock';
